explicit typecasts for inspect outputs

+++ stdlib/inspect/query_row_count_estimate.sql
@@ -1,8 +1,8 @@
-create or replace function @extschema@.query_row_count_estimate(query text) returns int
+create or replace function @extschema@.query_row_count_estimate(query text) returns bigint
 as $$
     declare
         explain_row record;
-        n_rows integer;
+        n_rows bigint;
     begin
         for explain_row in execute 'explain ' || query loop
             n_rows := substring(explain_row."QUERY PLAN" FROM ' rows=([[:digit:]]+)');
